# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Lost & Found API
  version: 2026.07.26
  description: Privacy-aware SEUSL lost-and-found API. AI scores are advisory and not proof of ownership.
servers:
  - url: /api
security:
  - cookieAuth: []
...
paths:
...
  /lost-items/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {tags: [Reports], security: [], responses: {'200': {description: Privacy-reduced report}}}
    put: {tags: [Reports], requestBody: {$ref: '#/components/requestBodies/MultipartReport'}, responses: {'200': {description: Owner/admin report update}}}
    delete: {tags: [Reports], responses: {'200': {description: Soft deleted}}}
...
  /found-items/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {tags: [Reports], security: [], responses: {'200': {description: Privacy-reduced report}}}
    put: {tags: [Reports], requestBody: {$ref: '#/components/requestBodies/MultipartReport'}, responses: {'200': {description: Owner/admin report update}}}
    delete: {tags: [Reports], responses: {'200': {description: Soft deleted}}}
...
components:
  securitySchemes:
    cookieAuth: {type: apiKey, in: cookie, name: lf_access}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-fA-F0-9]{24}$'}}
...
  requestBodies:
...
    MultipartReport:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              itemName: {type: string, maxLength: 150}
              category: {type: string}
              description: {type: string, minLength: 10, maxLength: 2000}
              images: {type: array, maxItems: 5, items: {type: string, format: binary}}
            required: [itemName, category, description]
...